big restructure. More code reuse in benches/examples, removed sandbox folder
