+++ .github/workflows/benchmark.yml
@@ -15,6 +15,5 @@ jobs:
 
       - uses: boa-dev/criterion-compare-action@v3
         with:
-          cwd: "./sled"
           # Needed. The name of the branch to compare with. This default uses the branch which is being pulled against
           branchName: ${{ github.base_ref }}
