+++ benches/resources/drivers.rs
@@ -1,0 +1,6 @@
+// Just rexport drivers from the examples folder
+
+#[path = "../../examples/drivers/mod.rs"]
+mod drivers;
+
+pub use drivers::*;
