+++ examples/resources/tui.rs
@@ -128,7 +128,7 @@ where
 
 #[allow(dead_code)]
 fn main() -> io::Result<()> {
-    let sled = Sled::new("./examples/config.toml").unwrap();
+    let sled = Sled::new("./examples/resources/config.toml").unwrap();
 
     let mut display = SledTerminalDisplay::start("Sled Visualizer", sled.domain());
     display.leds = sled.read_colors_and_positions();
