--- sandbox/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "sandbox"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-sled = { path = "../sled" }
-colored = "2.0.4"
-bevy = { version = "0.12", features = ["dynamic_linking"] }
-
-[workspace]
-resolver = "2"
-
-[profile.dev]
-opt-level = 1
-
-[profile.dev.package."*"]
-opt-level = 3
