+++ src/lib.rs
@@ -26,7 +26,7 @@
 //! Set all vertices to white:
 //! ```rust
 //!# use sled::{Sled, color::Rgb};
-//!# let mut sled = Sled::new("./examples/config.toml").unwrap();
+//!# let mut sled = Sled::new("./examples/resources/config.toml").unwrap();
 //! sled.set_vertices(Rgb::new(1.0, 1.0, 1.0));
 //! ```
 //!
@@ -34,7 +34,7 @@
 //! ```rust
 //!# use sled::{Sled, SledError, color::Rgb};
 //!# fn main() -> Result<(), SledError> {
-//!# let mut sled = Sled::new("./examples/config.toml").unwrap();
+//!# let mut sled = Sled::new("./examples/resources/config.toml").unwrap();
 //! sled.set_at_dist(1.0, Rgb::new(1.0, 0.0, 0.0))?;
 //!# Ok(())}
 //! ```
@@ -42,7 +42,7 @@
 //! Set each LED using a function of its direction from point `(2, 1)`:
 //!```rust
 //!# use sled::{Sled, Vec2, color::Rgb};
-//!# let mut sled = Sled::new("./examples/config.toml").unwrap();
+//!# let mut sled = Sled::new("./examples/resources/config.toml").unwrap();
 //! sled.map_by_dir_from(Vec2::new(2.0, 1.0), |dir| {
 //!     let red = (dir.x + 1.0) * 0.5;
 //!     let green = (dir.y + 1.0) * 0.5;
@@ -54,7 +54,7 @@
 //! ```rust
 //!# use sled::{Sled, SledError};
 //!# fn main() -> Result<(), SledError> {
-//!# let mut sled = Sled::new("./examples/config.toml")?;
+//!# let mut sled = Sled::new("./examples/resources/config.toml")?;
 //! sled.modulate_segment(2, |led| led.color * 0.5)?;
 //!# Ok(())
 //!# }
@@ -63,7 +63,7 @@
 //! ```rust
 //!# use sled::{Sled, SledError, Vec2, Filter, color::Rgb};
 //!# fn main() -> Result<(), SledError> {
-//!# let mut sled = Sled::new("./examples/config.toml")?;
+//!# let mut sled = Sled::new("./examples/resources/config.toml")?;
 //! let circle_1: Filter = sled.get_within_dist_from(
 //!     2.0,
 //!     Vec2::new(-0.5, 0.0)
@@ -85,7 +85,7 @@
 //! Once you're ready to display these colors, you'll probably want them packed in a nice contiguous array of RGB values. There are a few methods available to pack germane data.
 //! ```rust
 //!# use sled::{Sled, Vec2, color::Rgb};
-//!# let sled = Sled::new("./examples/config.toml").unwrap();
+//!# let sled = Sled::new("./examples/resources/config.toml").unwrap();
 //! let colors_f32: Vec<Rgb> = sled.read_colors();
 //! let colors_u8: Vec<Rgb<_, u8>> = sled.read_colors();
 //! 
@@ -97,7 +97,7 @@
 //! Or, if you just want the [Led] structs raw...
 //! ```rust
 //!# use sled::{Sled, Led};
-//!# let sled = Sled::new("./examples/config.toml").unwrap();
+//!# let sled = Sled::new("./examples/resources/config.toml").unwrap();
 //! let leds: Vec<Led> = sled.read();
 //! ```
