+++ src/sled/indexical.rs
@@ -19,7 +19,7 @@ impl Sled {
     /// ```rust
     ///# use sled::{Sled, SledError, color::Rgb};
     ///# fn demo() -> Result<(), SledError> {
-    ///# let mut sled = Sled::new("./examples/config.toml")?;
+    ///# let mut sled = Sled::new("./examples/resources/config.toml")?;
     /// sled.modulate(0,
     ///     |led| led.color + Rgb::new(0.5, 0.0, 0.0)
     /// )?;
@@ -61,7 +61,7 @@ impl Sled {
     /// For each method granting mutable access to each LED in the system.
     /// ```rust
     ///# use sled::{Sled, color::Rgb};
-    ///# let mut sled = Sled::new("./examples/config.toml").unwrap();
+    ///# let mut sled = Sled::new("./examples/resources/config.toml").unwrap();
     /// sled.for_each(|led| {
     ///     if led.index() % 2 == 1 {
     ///         led.color = Rgb::new(1.0, 0.0, 0.0);
@@ -95,7 +95,7 @@ impl Sled {
     /// ```rust
     ///# use sled::{Sled, SledError};
     ///# fn demo() -> Result<(), SledError> {
-    ///# let mut sled = Sled::new("./examples/config.toml")?;
+    ///# let mut sled = Sled::new("./examples/resources/config.toml")?;
     /// sled.modulate_range(0..50, |led| led.color * 0.5)?;
     ///# Ok(())
     ///# }
@@ -134,7 +134,7 @@ impl Sled {
     /// For each method granting mutable access to each LED with an index in `index_range`
     /// ```rust
     ///# use sled::{Sled, color::Rgb};
-    ///# let mut sled = Sled::new("./examples/config.toml").unwrap();
+    ///# let mut sled = Sled::new("./examples/resources/config.toml").unwrap();
     /// sled.for_each_in_range(50..100, |led| {
     ///     if led.index() % 2 == 1 {
     ///         led.color = Rgb::new(1.0, 0.0, 0.0);
